arch-review: extract routes_node.py and add api_error helper
Move the 7 node_* handlers and the 3 package/device handlers out of
routes_api.py into ciao/web/routes_node.py (the multi-device host/client
subsystem plus the package status/update surface, which app.py registers
twice as /api/package/* and the never-proxied /api/device/* twins). Adds
api_error(msg, status) to routes_helpers.py matching the hand-built
{"error": ...} envelope, used in the new file only. Moves the shared
_parse_set_cookie_session helper to routes_helpers.py so routes_node,
routes_auth, and node_proxy all import it from the helper module. Drops the
now-orphaned package_version and network_addresses imports from
routes_api.py and repoints test imports/patch targets at routes_node.
